requirements.txt
chromadb>=0.5.0
openai>=1.40.0
requests>=2.32.0
tqdm>=4.66.0
python-dotenv>=1.0.1
rich>=13.7.0
typer>=0.12.3
